refactor homepage into homepage sections
