+++ frontend/src/components/Header.tsx
@@ -1,27 +1,37 @@
 "use client";
 
-import clsx from "clsx";
-import GrantsLogo from "public/img/grants-logo.svg";
-import { useFeatureFlags } from "src/hooks/useFeatureFlags";
-import { useSnackbar } from "src/hooks/useSnackbar";
-import { useUser } from "src/services/auth/useUser";
-import { IndexType } from "src/types/generalTypes";
-import { isCurrentPath, isExternalLink } from "src/utils/generalUtils";
+import {
+  Suspense,
+  useCallback,
+  useEffect,
+  useMemo,
+  useState,
+} from "react";
 
+import clsx from "clsx";
 import { useTranslations } from "next-intl";
 import Image from "next/image";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
-import { Suspense, useCallback, useEffect, useMemo, useState } from "react";
+import GrantsLogo from "public/img/grants-logo.svg";
+import { USWDSIcon } from "src/components/USWDSIcon";
+import { useFeatureFlags } from "src/hooks/useFeatureFlags";
+import { useSnackbar } from "src/hooks/useSnackbar";
+import { useUser } from "src/services/auth/useUser";
+import { IndexType } from "src/types/generalTypes";
+import {
+  isCurrentPath,
+  isExternalLink,
+} from "src/utils/generalUtils";
+
 import {
   GovBanner,
+  Header as USWDSHeader,
   NavMenuButton,
   PrimaryNav,
   Title,
-  Header as USWDSHeader,
 } from "@trussworks/react-uswds";
 
-import { USWDSIcon } from "src/components/USWDSIcon";
 import NavDropdown from "./NavDropdown";
 import { RouteChangeWatcher } from "./RouteChangeWatcher";
 import { UserControl } from "./user/UserControl";
