+++ frontend/src/components/homepage/IconInfoSection.tsx
@@ -1,6 +1,5 @@
-import { UswdsIconNames } from "src/types/generalTypes";
-
 import { USWDSIcon } from "src/components/USWDSIcon";
+import { UswdsIconNames } from "src/types/generalTypes";
 
 type Props = {
   description: string;
