+++ frontend/src/components/homepage/sections/HomepageBuilding.tsx
