+++ frontend/src/components/homepage/sections/HomepageInvolved.tsx
@@ -1,0 +1,54 @@
+import { useTranslations } from "next-intl";
+import HomePageSection from "src/components/homepage/homePageSection";
+
+import { Grid } from "@trussworks/react-uswds";
+
+import IconInfo from "../IconInfoSection";
+
+const codeLink = "https://wiki.simpler.grants.gov/get-involved/github-code";
+const discourseLink = "https://simplergrants.discourse.group/";
+const participateLink = "https://ethn.io/91822";
+
+const InvolvedContent = () => {
+  const t = useTranslations("Involved");
+
+  return (
+    <HomePageSection
+      className="padding-y-6 bg-base-lightest"
+      data-testid="homepage-involved"
+      title={t("title")}
+    >
+      <Grid row className="padding-y-2" gap="md">
+        <Grid col={6}>
+          <IconInfo
+            description={t("technicalDescription")}
+            iconName="code"
+            link={codeLink}
+            linkText={t("technicalLink")}
+            title={t("technicalTitle")}
+          />
+          <br />
+          <a
+            href={discourseLink}
+            className="font-sans-md line-height-sans-4"
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            {t("discourseLink")}
+          </a>
+        </Grid>
+        <Grid col={6}>
+          <IconInfo
+            description={t("participateDescription")}
+            iconName="chat"
+            link={participateLink}
+            linkText={t("participateLink")}
+            title={t("participateTitle")}
+          />
+        </Grid>
+      </Grid>
+    </HomePageSection>
+  );
+};
+
+export default InvolvedContent;
