+++ frontend/src/components/vision/sections/VisionMission.tsx
@@ -1,5 +1,7 @@
-import { useMessages, useTranslations } from "next-intl";
-
+import {
+  useMessages,
+  useTranslations,
+} from "next-intl";
 import { USWDSIcon } from "src/components/USWDSIcon";
 import VisionPageSection from "src/components/vision/visionPageSection";
